refactor(tests): move ported tests into more appropriate file
